+++ cognite/experimental/_api/transformations.py
@@ -62,12 +62,18 @@ class TransformationsAPI(APIClient):
         utils._auxiliary.assert_type(transformation, "transformation", [Transformation, list])
         return self._create_multiple(transformation)
 
-    def delete(self, id: Union[int, List[int]] = None, external_id: Union[str, List[str]] = None) -> None:
+    def delete(
+        self,
+        id: Union[int, List[int]] = None,
+        external_id: Union[str, List[str]] = None,
+        ignore_unknown_ids: bool = False,
+    ) -> None:
         """`Delete one or more transformations. <https://docs.cognite.com/api/playground/#operation/deleteTransformations>`_
 
         Args:
             id (Union[int, List[int]): Id or list of ids.
             external_id (Union[str, List[str]]): External ID or list of external ids.
+            ignore_unknown_ids (bool): Ignore IDs and external IDs that are not found rather than throw an exception.
 
         Returns:
             None
@@ -80,7 +86,7 @@ class TransformationsAPI(APIClient):
                 >>> c = CogniteClient()
                 >>> c.transformations.delete(id=[1,2,3], external_id="function3")
         """
-        self._delete_multiple(ids=id, external_ids=external_id, wrap_ids=True)
+        self._delete_multiple(ids=id, external_ids=external_id, wrap_ids=True, ignore_unknown_ids=ignore_unknown_ids)
 
     def list(self, include_public: bool = True, limit: Optional[int] = LIST_LIMIT_DEFAULT,) -> TransformationList:
         """`List all transformations. <https://docs.cognite.com/api/playground/#operation/transformations>`_
@@ -141,8 +147,9 @@ class TransformationsAPI(APIClient):
         """`Retrieve multiple transformations. <https://docs.cognite.com/api/playground/#operation/getTransformation>`_
 
         Args:
-            ids (List[int]): List of ids to retrieve
-            external_ids (List[str]): List of external ids to retrieve
+            ids (List[int]): List of ids to retrieve.
+            external_ids (List[str]): List of external ids to retrieve.
+            ignore_unknown_ids (bool): Ignore IDs and external IDs that are not found rather than throw an exception.
 
         Returns:
             TransformationList: Requested transformation or None if it does not exist.